rename database to db
